Lower the starting cost of scale to $1k/month
