+++ src/components/Pricing/PricingSlider/LogSlider.tsx
