+++ src/components/Pricing/PricingSlider/index.tsx
