+++ src/components/Pricing/PricingSlider/log-slider.scss
