+++ src/components/Pricing/PricingTable/Plans.js
@@ -2,51 +2,9 @@ import { TrackedCTA } from 'components/CallToAction/index.js'
 import { Cohorts, FeatureFlags, Funnels, PathAnalysis, SessionRecordings } from 'components/Icons/Icons'
 import Link from 'components/Link'
 import React from 'react'
+import { SCALE_MINIMUM_PRICING } from '../constants'
 import { Features, Plan, Price, Section } from './Plan'
-
-export const features = {
-    Platform: [
-        { title: 'Funnels & trends', icon: <Funnels className="w-3" /> },
-        { title: 'Feature flags', icon: <FeatureFlags className="w-3" /> },
-        { title: 'Cohorts & retention', icon: <Cohorts className="w-3" /> },
-        { title: 'Session recordings', icon: <SessionRecordings className="w-3" /> },
-        { title: 'Path analysis', icon: <PathAnalysis className="w-3" /> },
-    ],
-    'Platform features': [
-        { title: 'Dashboards' },
-        { title: 'Event autocapture' },
-        { title: 'Annotations' },
-        { title: 'API' },
-        { title: 'Plugins' },
-        { title: 'Data I/O' },
-    ],
-    'Advanced features': [
-        { title: 'Correlation analysis' },
-        { title: 'Priority Session Recordings' },
-        { title: 'Multivariate testing' },
-        { title: 'Deep Dive Dashboards' },
-        { title: 'Multiple projects' },
-    ],
-    Collaboration: [
-        { title: 'User permissions' },
-        { title: 'Dashboard collaboration tools' },
-        { title: 'Team event management tools' },
-    ],
-    'Account & support': [
-        { title: 'Team training' },
-        { title: 'Dashboard configuration support' },
-        { title: 'Monitoring setup support' },
-        { title: 'Infrastructure management pairing' },
-        { title: 'SLA w/ downtime developer pairing' },
-    ],
-    'Ops & security': [
-        { title: 'Project permissions' },
-        { title: 'SSO, SAML' },
-        { title: 'Configurable backups' },
-        { title: 'Instance monitoring' },
-    ],
-    Benefits: [{ title: 'Hosted by PostHog' }, { title: 'Automatic upgrades' }, { title: 'Community Slack support' }],
-}
+import { features } from '../constants'
 
 export const OpenSource = () => {
     return (
@@ -99,7 +57,8 @@ export const Scale = ({
                     <Section title="Pricing starts at" className="mt-auto">
                         <div className="flex justify-between items-center">
                             <Price>
-                                $2,000<span className="text-base opacity-50">/mo</span>
+                                ${Math.round(SCALE_MINIMUM_PRICING).toLocaleString()}
+                                <span className="text-base opacity-50">/mo</span>
                             </Price>
                             {!hideCalculator && (
                                 <Link className="text-yellow font-bold" onClick={() => setOpen(true)}>
